refactor : edit expression : unsigned int -> size_t

+++ ft_strlcpy.c
@@ -6,7 +6,7 @@
 /*   By: rjeong <rjeong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/07 22:17:10 by rjeong            #+#    #+#             */
-/*   Updated: 2022/07/08 22:51:14 by rjeong           ###   ########.fr       */
+/*   Updated: 2022/07/08 23:29:44 by rjeong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlcpy(char *dst, const char *src, size_t dstsize)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	while (i < dstsize - 1)
